Clean up Rational/Polynomial methods, pull out to prototype.
This should be a better approach, because it means there doesn't
need to be a closure for every call to the constructor,
and all of the methods are created once and shared on the prototype.

Additionally clean up lots of things and add tests.

Follow a Z/Q/R convention for argument types.